Addition of __name__ function for code shrinking (hudge line/ job ratio impact)
